Move the add cases modal considering its no longer only on the multi-facility-page
