+++ src/components/AddCasesModal/AddCasesModalContent.tsx
@@ -4,18 +4,18 @@ import numeral from "numeral";
 import React, { useCallback } from "react";
 import styled from "styled-components";
 
-import Colors from "../../components/design-system/Colors";
-import { DateMMMMdyyyy } from "../../components/design-system/DateFormats";
-import InputButton from "../../components/design-system/InputButton";
-import InputDate from "../../components/design-system/InputDate";
-import Tooltip from "../../components/design-system/Tooltip";
+import { ModelInputs } from "../../page-multi-facility/types";
+import Colors from "../design-system/Colors";
+import { DateMMMMdyyyy } from "../design-system/DateFormats";
+import InputButton from "../design-system/InputButton";
+import InputDate from "../design-system/InputDate";
+import Tooltip from "../design-system/Tooltip";
 import {
   getTotalPopulation,
   ModelInputsPopulationBrackets,
   totalConfirmedCases,
-} from "../../components/impact-dashboard/EpidemicModelContext";
-import { AgeGroupGrid } from "../../components/impact-dashboard/FacilityInformation";
-import { ModelInputs } from "../types";
+} from "../impact-dashboard/EpidemicModelContext";
+import { AgeGroupGrid } from "../impact-dashboard/FacilityInformation";
 
 const noDataColor = Colors.darkRed;
