+++ src/components/rt-timeseries/RtTimeseriesContainer.tsx
@@ -5,8 +5,8 @@ import styled from "styled-components";
 import { FacilityContext } from "../../contexts/facility-context/FacilityContext";
 import { RtData } from "../../infection-model/rt";
 import { updateFacilityRtData } from "../../infection-model/rt";
-import AddCasesModal from "../../page-multi-facility/AddCasesModal";
 import { Facility } from "../../page-multi-facility/types";
+import AddCasesModal from "../AddCasesModal";
 import Colors from "../design-system/Colors";
 import HelpButtonWithTooltip from "../design-system/HelpButtonWithTooltip";
 import Loading from "../design-system/Loading";
